Inline and clean lipo code
